2026-03-22 添加修改Name函数,修复接受密码为空的问题,优化分层逻辑.虚了虚了

+++ internal/middleware/auth.go
@@ -48,7 +48,7 @@ func (m *JWTMiddleware) Auth() gin.HandlerFunc {
 		}
 
 		// 将用户信息存入上下文，供后续handler使用
-		c.Set("user_id", claims.UserID)
+		c.Set("user_id", uint(claims.UserID))
 		c.Set("account", claims.Account)
 		c.Next()
 	}
@@ -85,7 +85,7 @@ func (m *JWTMiddleware) OptionalAuth() gin.HandlerFunc {
 		}
 
 		// token有效，将用户信息存入上下文
-		c.Set("user_id", claims.UserID)
+		c.Set("user_id", uint(claims.UserID))
 		c.Set("account", claims.Account)
 		c.Next()
 	}
